:bomb:💾 SaveGame saves a list of BombDTOs

+++ src/Swarm.Application/DTOs/BombDTO.cs
@@ -1,3 +1,3 @@
 namespace Swarm.Application.DTOs;
 
-public readonly record struct BombDTO(string Identifier, int CooldownSeconds);
+public readonly record struct BombDTO(string Identifier, float CooldownSeconds);
